+++ packages/gamut-labs/src/ListSection/index.tsx
@@ -1,6 +1,7 @@
 import { theme } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
-import React, { Children, useState } from 'react';
+import { Children, useState } from 'react';
+import * as React from 'react';
 
 import { PageSection, SectionButton } from '..';