+++ packages/gamut-labs/src/TabsExperimental/Tab.tsx
@@ -1,7 +1,7 @@
 import { Box, TextButton } from '@codecademy/gamut';
 import { css, states, useCurrentMode } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
-import React from 'react';
+import * as React from 'react';
 
 import { TabProps } from './types';