+++ packages/variance/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.20.5](https://github.com/Codecademy/gamut/compare/@codecademy/variance@0.20.4...@codecademy/variance@0.20.5) (2022-12-06)
+
+**Note:** Version bump only for package @codecademy/variance
+
+
+
+
+
 ### [0.20.4](https://github.com/Codecademy/gamut/compare/@codecademy/variance@0.20.3...@codecademy/variance@0.20.4) (2022-11-29)
 
 **Note:** Version bump only for package @codecademy/variance